refactor(combat): lift shared combat types to src/game/combat/types.ts
Moves CombatState, BossIntroStage, BOSS_INTRO_STAGE_MS, Enemy, and
derivePhase out of useCombatLoop into a shared module. Resolves the
type-only cyclic import in useCombatTick (was reaching back into a
sibling hook for the Enemy type).

useCombatLoop re-exports the types so external consumers (CombatScene,
MonsterTooltip, world.tsx) don't need to change. The orchestrator drops
from 433 to 385 lines, under the 400-line target from the plan.

Also: biome formatting pass + the spawn-reset effect's dep refactored
into a captured `enemyScaled` variable so the useExhaustiveDependencies
rule passes without losing the "fire-on-scaled-identity" semantics.
